requirements.txt
matplotlib>=3.7.1
seaborn>=0.12.2
pandas>=2.0.0
numpy>=1.24.3
PyQt6>=6.4.2
PyQt6-Qt6>=6.4.2
PyQt6-sip>=13.4.1
python-dotenv>=1.0.0
SpeechRecognition>=3.10.0
pyttsx3>=2.90
gTTS>=2.3.2
google-generativeai>=0.3.0
pyaudio>=0.2.13
pvporcupine>=2.2.0
webrtcvad>=2.0.10
pygame>=2.5.0
opencv-python>=4.8.0
Pillow>=10.0.0 
serial